fix move stuff around
