+++ tests/test_parts.py
@@ -1,9 +1,7 @@
 import unittest
 
-from semver import Version
-
 import context
-from versions import parse_version, is_compatible
+from part_tables import is_compatible
 
 context.unused_import_dummy = 1
 
@@ -14,22 +12,6 @@ def get_row(first: str, last: str, active: bool):
         'lastUpdated': last,
         'status': 'active' if active else 'depreciated',
     }
-
-
-class TestParseVersion(unittest.TestCase):
-    def test_default(self):
-        default = Version(major=123)
-        self.assertEqual(parse_version('', default=default), default)
-        with self.assertRaises(ValueError):
-            parse_version('')
-
-    def test_incomplete(self):
-        self.assertEqual(parse_version('1'), '1.0.0')
-        self.assertEqual(parse_version('2.0'), '2.0.0')
-        self.assertEqual(parse_version('3.0.'), '3.0.0')
-
-    def test_garbage(self):
-        self.assertEqual(parse_version('1.2.3garbage'), '1.2.3')
 
 
 class TestValidateVersion(unittest.TestCase):
